Retire utils special-char logic to canonical corpus CLI
